// src/Types/DailyStatistics/IDailyStatistics.ts
export interface IDailyStatistics {
  amount: number;
  avgSalary: number;
  createdAt: string;
  leastSeniorityOffers: string;
  leastSeniorityOffersCount: number;
  mostSeniorityOffers: string;
  mostSeniorityOffersCount: number;
  secondSeniorityOffers: string;
  secondSeniorityOffersCount: number;
  type: string;
}